+++ packages/grpc-js-core/test/test-channel-credentials.ts
@@ -36,7 +36,7 @@ const readFile: (...args: any[]) => Promise<Buffer> = promisify(fs.readFile);
 // A promise which resolves to loaded files in the form { ca, key, cert }
 const pFixtures = Promise
                       .all(['ca.pem', 'server1.key', 'server1.pem'].map(
-                          (file) => readFile(`test/fixtures/${file}`)))
+                          (file) => readFile(`${__dirname}/fixtures/${file}`)))
                       .then((result) => {
                         return {ca: result[0], key: result[1], cert: result[2]};
                       });
